@@ -1,21 +1,18 @@
 [tool.poetry]
 name = "xtrain"
-version = "0.1.3"
+version = "0.0.0"
 description = "A Flax trainer"
 authors = ["Ji Yu <jyu@uchc.edu>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.9 <3.11"
+python = "^3.10"
 flax = "^0.7.2"
-
-[tool.poetry.group.dev]
-optional = true
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
-jax = {extras = ["cuda11-pip"], version = "^0.4.14"}
-tensorflow-cpu = "^2.13.0"
+jax = {extras = ["cuda11_pip"], version = "^0.4, <=0.4.24"}
+nvidia-cudnn-cu11 = "^8.0" # bypass jax packaging error
 
 [tool.poetry.group.docs]
 optional = true
@@ -32,5 +29,4 @@ build-backend = "poetry.core.masonry.api"
 [[tool.poetry.source]]
 name = "jax"
 url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
-default = false
-secondary = false
+priority = "supplemental"
